Update patch set 3
Patch Set 3:

Tested with a netvirt and genius patch [0]. This patch should make it easier for folks to use multipatch without having to require integration. It can be triggered from any Gerrit patch and the list of patches need to be provided.

An improvement probably for another patch is to automatically assume the current patch is the patch that should be built.


https://logs.opendaylight.org/releng/jenkins092/integration-multipatch-test-carbon/203/

Patch-set: 3