Update patch set 3
Patch Set 3:

> > a few things to understand/figure-out:
 > >
 > > 1) I don't see a single log.html get created. Can we still get
 > > this?
 > Do we need one or can that be figured in a later patch? The smaller
 > log.htmls load faster. It is a pain if wanting a quick look at what
 > failed all in one shot.

it's not necessary, but I don't know how easy it will be to know
exactly which .html to click on to go see the failures. It's also
not going to let us easily navigate the order of suites. There are
times when I'm drilling back through previous suites looking for
some info (like a mac address).

 > >
 > > 2) the resulting output.xml in the robot-plugin folder is not as
 > > big
 > > as it is when we are seeing the normal pybot process choke and
 > > get killed. That's the condition we need to see if this patch
 > > resolves. I think it happens when that file approaches 54M
 > All the files aggregated should end up about the same so not sure
 > why we need to know this. As long as each file is less than 54M
 > then we are good right? And that is what this patch does.

my point was maybe this sandbox job was one of the few that didn't
have large enough log size(s) to effectively choke our VM building
them. I wanted to make sure we weren't doing all this work to not
even resolve our issues.

 > >
 > > 3) If test case names must be unique we need to find a process
 > > to guarantee that (non human process) because it's pretty
 > > easy to write a new test suite and one of the test cases is
 > > "Create Networks"
 > >
 > > 4) Is the jenkins robot plugin just parsing all of the individual
 > > robot.html files to know the pass/fail count? that's cool, if
 > > so.
 > Looks like it since it had all the tests.

yeah, but because of 3 we don't get a fully accurate view.

Patch-set: 3