Upgrade robot vms to use 4G RAM from 2G
some robot jobs are creating log files (output.xml) in excess
of 1.5G in size. When robot parses that file to generate it's
html reports it reads the entire file (I believe) in to memory
which ends up causing an OOM scenario in a 2G system. For now,
this will help, but as jobs increase in size and duration,
even 4G will not be enough.

Change-Id: I71c72a8ad62a05549fc4623531f4c7b9387505bc
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>