Update patch set 5
Patch Set 5:

> thanks for this. looks like the sandbox run shows it takes aprox
 > 10m for rabbitmq to be up. just fyi.

agreed - about 10 minutes. But also doesn't seem to matter as the computes still finished stacking before the control node. Looks like the control node still needed a couple more minutes to stack. So the overall time should be the same since it is the control node that is the longest.

The previous workaround had an arbitrary 6m wait - which was useless since the issue is really a bug in the nova code, the workaround was just getting lucky.

Patch-set: 5
Label: Verified=0