Use grep -E to grab relevant exception logs
There is some quirk/bug/problem with sed that it's not working like
we expect to collect the Exception logs. Sometimes it would not
return the expected output. There is probably the right way to do
it but I did not find it. I was trying to match on Exception$ which
would match our "java.lang.NullPointerException" which always
ends with a newline. We were not collecting it before.

Change-Id: I5c8f5bbae2ec44198d53cfc84ead779b3e280cf2
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>