Update patch set 1
Patch Set 1:

Jamo, yes, this should be fine to merge. Genius code will not act on this because genius auto-tz will only be enabled in master/oxygen. So in carbon and nitrogen jobs, this config will not be acted on by genius code, since genius auto-tz is not enabled there.

Patch-set: 1