Update packer config to work in Vexxhost
Improved the templates to support SOCKS 5 proxy configuration
via the `ssh_proxy_host` configuration that is optional. This
can be configured in the vars file.

Additionally we are now prefixing image names with "ZZCI - " so that
you can immediately tell from the name that the image is managed by
our automated systems as well as having them sorted last. This should
make it easier to manage images in the WebUI.

Update packer version from 0.12.2 to 1.1.3 so that we can use the
ssh_proxy_host parameter.

Issue: RELENG-706
Change-Id: I3b7200c6a917a961f9da38e5eeea95347ceed4a7
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>