Update patch set 1
Patch Set 1:

> an HOUR!?!?!

just to be cautious... We were waiting 30m, but the infra issue with the rpms takes about 20m sometimes, so that would only leave 10 minutes for rabbitmq to be ready. The control node averaged out to 17m before rabbitmq was ready. So the rabbitmq check was failing. With 60m we should finish really just after 30m. And ideally the infra should be fixed so that the repo mirrors stuff finishes quickly - which it does in the passing jobs.

Patch-set: 1