Fix HEAT userdata scripts
The migration of the Gerrit server to a different environment has caused
the startup scripts to fail. The issue is that the new location does not
allow a ping. Instead, we will use the host command to do a DNS query
since that tests for network connetivity just as well.

Change-Id: I83755e7fca0afda7e8d71fa9598dc73b7db94853
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>