Always keep copy of RPM in build out dir
The RelEng and sandbox silos need to behave differently with regard to
uploading artifacts to Nexus (sandbox shouldn't, RelEng should).
However, moving the artifacts in RelEng but not in the sandbox was
causing the new dependency tests to work in the sandbox but not in
RelEng because the artifact no longer existed at the path we expected.
Instead of moving the artifacts, copy them.

We needed to mv and not cp initially because we were still hosting on
Jenkins and after a cp it found both old and new, but we no longer have
that logic.

Change-Id: I22ac9cca0c9a5a4ff9d3a11b841a6584ffeb3aa5
Fixes: INTPAK-119
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>