Do RPM tests in RPM build job to gate publish
We currently only run our RPM tests after the RPM build job finishes.
This isn't good because by the time we've tested the RPM it has already
been pushed to Nexus and is being consumed by downstreams. Instead we
should fail the RPM build job if the tests fail and not publish.

Change-Id: I215de288923a308b95ebd242578d2a90d0b8e6ec
Fixes: INTPAK-115
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>