Use right pid for jstack collection
The jstack collection fails in post-carbon distros.
Starting with karaf 4 there are two processes for karaf
where the existing code expected one.

Grep was reworked to use the main process of the
karaf shell to return a single pid. This works for
carbon also since that is the only process found.

Change-Id: Ic3b0fc0f3c171e5c18228e3e537c79983dc26f96
Signed-off-by: Sam Hague <shague@redhat.com>