Add a job to cleanup orphaned nodes
This job assumes that releng minions have a 'prd' prefix and
sandbox has a 'snd' prefix. These minions are Jenkins managed
minions and are not Heat managed minions.

The script get's a list of jobs with the prd or snd prefix and then
checks against Jenkins to see if the nodes are attached to Jenkins
if they are not then issue a delete call as it is likely Jenkins
lost track of them.

Issue: RELENG-724
Change-Id: I624102ace279d49d2696d287cfea9e4e89ba7a20
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>