Update HEAT scripts to use local gerrit mirror
With Gerrit no longer living in the same cloud as the CI infrastructure
a localized mirror has been put in place to speed up git operations as
much as possible. Since instance bring-ups should be as fast as we can
possibly make them we need to point at this localized mirror for the
spin-up scripts.

Change-Id: I7bbeaf36bd7ccd5f5c9d45707d3c1f8a8507ee5f
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>