Add jenkins param to use featuresBoot (or not)
This should only add a parameter and default everything to
'True' so there will be no functional changes to the jobs.

This is to prepare for the next step which will stop
the netvirt csit from installing with featuresBoot in
order to let those jobs install from the karaf console.

Change-Id: I1ed9c28b3360a14286c6d512cabe3db7a696e531
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>