Improve Stack cleanup script to check on each delete
There is a race condition between when we query Jenkins for a list
of in use builds to compare with stacks and when they are actually
deleted. Shorten that race condition by querying on every stack
check.

Hopefully this will reduce the chance that Jenkins deletes a stack
that is in progress building and has not yet attached to Jenkins.

Issue: RELENG-75
Change-Id: I2fcc0a9bef34708062e69f4d8a4ee04dfd6108b5
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>