Add check for max robot node types
We do not allow more than 1 robot node type to be in use at the same
time. This ensures that we can properly limit the number of
simultanious robot builds properly due to the way the Jenkins
OpenStack cloud plugin works. All templates should always use the
same Robot VM so any updates to change this needs to be changed in
all robot templates.

Change-Id: Id2087bb92a860379fc2437a1a58cac578e853be5
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>