Update patch set 5
Patch Set 5: Code-Review+1

LGMT, but one point: The whole --no-auto-refresh hand stand, as well as possible even the fact that you couldn't use featuresBoot to start with, is only because of CONTROLLER-1764... my understanding was that skitt actually fixed that (but he should confirm) so it may be possible to do this more easily.  However, as you say that this also works, I've no objection to this going in already anyway - perhaps only to be simplified again subsequently.

Patch-set: 5
Reviewer: Gerrit User 5867 <5867@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1