Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Is this because we hardcode what image to use as below? Starting with pike we went to the -disk. Maybe the -uec are still there and can be used?
> 
> ${CIRROS_stable/newton}    cirros-0.3.4-x86_64-uec
> ${CIRROS_stable/ocata}    cirros-0.3.4-x86_64-uec
> ${CIRROS_stable/pike}    cirros-0.3.5-x86_64-disk
> ${CIRROS_master}    cirros-0.3.5-x86_64-disk


No Sam It seems devstack uses qemu-img command for something and it has been removed recently and replaced. It will take longer time to change in devstack. So as a quick fix, we can get the required image for Pike Tests and workaround the problem

Patch-set: 2
Label: Verified=0