Add job to copy logs from a sandbox path to prod
As requested by the netvirt project to provide capabilities for
storing logs for a longer amount of time in cases where they need
to perform complex troubleshooting. This patch adds a job which
when given a log path will pull the logs from path on sandbox
and copy them to the releng system.

Change-Id: I96c8b8eb8d54c266ccab1c9bf19450b11d5bea69
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>