Update patch set 3
Patch Set 3: Code-Review-1

> this looks right to me. I'm used to seeing Daniel +1 his patches
 > when
 > they are ready, so take my +2 and self-merge when ready.

You know me well Jamo:)

It seems right to me, but it's failing with an SSH timeout eventually. Not sure if the test is just taking too long or what, but not quite ready to merge. -1ing myself for now while I wait on more test results but it might actually be good.

Patch-set: 3
Label: Code-Review=-1