Update Merge-$STREAM views via AR jobs
`Merge-$(STREAM)` views lists all the merge jobs for the projects
participating in the release cycle. Update the views by getting
a list of all projects in autorelease $STREAM and create a new
yaml file as a part of the autorelease jobs.

JJB 2.0.5 is released, includes the relevent changes required
for manging views with JJB.

ISSUE: RELENG-76
Change-Id: I7d5e9d28bd6d3f41c54bcf17bb7350cffda3c545
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>