Update patch set 1
Patch Set 1:

Can we also split this into a separate script? It has a different purpose so I'd rather not mix the tasks into the same script. Make this a macro and call the macro in the job-template.

(We should probably turn the other thing into a macro too but that should be a separate task)

Patch-set: 1