Update patch set 1
Patch Set 1:

Right. I plan to push this out to every job in releng/builder. Just didn't want to pull in too many things in one patch. Let's get this rolled out and test it first just to see that it works and then we can make patches for the other templates later.

Patch-set: 1