Update patch set 1
Patch Set 1:

> this is going to cause job history loss, right? if so, we need to
 > discuss and maybe inform.

No it won't it's the same change as Iaf6f026853312098bf9bdcb562ad54450b8671fa and does not affect anyone. All the jobs will continue to exist as the same job. The proposed "{prefix}" field is empty so no job names are changed. It's only useful for the sandbox where you can edit the prefix and upload a job with your name.

Patch-set: 1