Switch the csit-3node-docker to use job prefix
Idea to add a new variable "{prefix}" to job templates. This
variable is defined already in defaults.yaml.
The prefix should always be blank in production thus producing the
same existing job names we are all used to. The prefix can be used
locally to add a prefix to job names when pushing jobs to the
sandbox. For example a prefix of "abc-" will be added to the front
of the job name allowing sandbox testers to be able to push a job
to the sandbox with a unique prefix to not overwrite other testers.

Additionally add an id inttest-csit-3node-docker to make the job
template more easy to maintain. Allows us to refactor the job
template name without having to update every single user of the
template every time as the id will always be the same.

Change-Id: I18d977a1a1377ac20b528b65d08b7c22fa888db9
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>