Add netvirt csit job for TDD and/or WIP code
We want to try having system tests drive development, but we
don't want to add failing test cases to existing jobs. The
existing jobs need to move and stay at 100% as we use them
to gate other dev patches. So, this new job can be a landing
spot for these new CSIT tests that may not be ready on the
feature dev side.

The job can be run with the patch test using the gerrit
keyword "test-netvirt-tdd-wip".

The job can also be a place for new CSIT that is almost
ready, but might need a place to bake to ensure we iron
out all the kinks. For these cases, we'll have to run the
job manually. Or we can set some sort of non frequent
schedule.

Change-Id: Iccdb1c0be8f859e7806b66e0c97e52c6e326d82c
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>