Update patch set 11
Patch Set 11: -Code-Review

ok, this is ready on my side.
sandbox:
https://jenkins.opendaylight.org/sandbox/job/jamo-common-func-netvirt-csit-1node-openstack-ocata-tdd-wip-oxygen/20/

I aborted sandbox because all I cared about was seeing the output of
'cat ${MEMCONF}' which looks right to me.

BUT!!! Sam worked on a slightly different solution to this problem.
his patch is here:

https://git.opendaylight.org/gerrit/c/67796/

pros and cons with each, I guess.

Patch-set: 11