Rename jdk to java-version
jdk is a keyword that is used for configuring the jdk configuration
in a job. This means all templates are inheriting and trying to set
this jdk regardless of if they jdk: '{jdk}' value in their configs
and appears to override that value incorrectly with 'openjdk8'
including the single quotes. This explains why many jobs are
printing the message: No JDK named ‘openjdk8’ found

Change-Id: I9bdf92bb759aec10f97ef355d4b04618cd58d2c8
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>