Add a swap file to build machines
We have heard this a few times now and again on the TSC mailing list
https://lists.opendaylight.org/pipermail/tsc/2018-February/008824.html

It might be beneficial to have some kind of swap available on the
system. There is a pretty good writeup about swap here:
https://chrisdown.name/2018/01/02/in-defence-of-swap.html

This was originally merged as Ifc9d57045b54e1b4f9943214e418b41fabf49d04
but then reverted to troubleshoot build issues. With CI stable now it is
worth trying again.

Change-Id: I85e491eb8944f3a672a11da9840275beadb104fa
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>