Add suse support for the snaps build
suse snaps rpm build and test happens inside a docker container
This patch requires the following patch to be merged:

https://git.opendaylight.org/gerrit/#/c/68114/

It reuses most of the building and testing scripts. Those scripts
for epel will be executed inside a VM (as it was done before) and
for suse, those will be executed inside a docker container.

It avoids duplicating the packaging-build-rpm-snap job using the
variable {distro} to differentiate between suse and epel. However,
to reuse most building and testing scripts, we inject the DISTRO
variable into the scripts to detect if we are inside the VM or the
container. When building a suse rpm, all scripts will be executed
twice and thus if we detect that it is a suse rpm and we are inside
a VM, it will exit the script

Change-Id: I31ee0558f3cc76e27f462f1d3ac528b5906a122c
Signed-off-by: Manuel Buil <mbuil@suse.com>