Update patch set 1
Patch Set 1:

It's because the name docker.yaml is already used in common-packer. This is to make a clear differentiation between the main docker.yaml and what is local configuration and not upstream.

The local-docker.yaml that lives in common-packer is a placeholder so that we can test it common-packer without having to checkout another project first. This way we can use the exact same docker.json template file in common-packer and in the local ci-management repository.

Patch-set: 1