Update patch set 2
Patch Set 2:

Agreed, it would good to test this change and see if it helps. 
Another thing to watch out if adding swap does not help with the overcommit tunables unless overcommit is set to 2 and a ratio is set, since the swap is only going to be used for page reclamation.

Patch-set: 2