Switch jjb-merge jobs to 4c-4g systems
With the addition of the Fluorine jobs the merge job now needs
more than 4g of RAM to actually run. Considering the number of
jobs ODL pushes bumping the CPU would save us some time with
uploads as well.

This patch breaks out the ci-jobs group so that we can individually
override configuration.

https://github.com/lfit/releng-global-jjb/blob/8e33e6ab798cd57a383cb8871d973de3fce6340c/jjb/lf-ci-jobs.yaml#L2-L9

Change-Id: I415aee470901fc3d69ea875efcae94db6aa6c9da
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>