Update patch set 35
Patch Set 35:

(4 comments)

> (6 comments)
 > 
 > > (7 comments)
 > >
 > > Thanks for the patch Manuel:)
 > >
 > > Can you show sandbox runs of all the affected jobs behaving the
 > > same, as well as the new jobs working?
 > 
 > I sandboxed epel-7 which is impacted for oxygen and nitrogen. What
 > else do you suggest?

What we typically do is provide sandbox logs for every job affected by a change for the patchset we think is ready to be merged. So in this case that would be quite a few jobs - all of the ones that touch these scripts/macros. It's a lot of jobs, but not really that abnormal, and I think very valuable.

Patch-set: 35