Update patch set 3
Patch Set 3:

> If all this job is doing is calling tox. Perhaps we should delete
 > the job-template and use the global-jjb provided tox job?
 > 
 > http://docs.releng.linuxfoundation.org/projects/global-jjb/en/latest/jjb/lf-python-jobs.html#tox-verify
 > 
 > This way you don't have to maintain a custom job.

That's a good idea Thanh.

The only thing is that I am not sure if we would like to bring jenkins_build_package.sh back any time soon. But if we do, it seems that it should be on a separate "release" specific job. But I am not aware about the possibility of project to setup custom release jobs. That script specifically uploads a package to pypi.

Patch-set: 3