Update patch set 3
Patch Set 3:

> The only thing is that I am not sure if we would like to bring
 > jenkins_build_package.sh back any time soon. But if we do, it seems
 > that it should be on a separate "release" specific job. But I am
 > not aware about the possibility of project to setup custom release
 > jobs. That script specifically uploads a package to pypi.

Similar to what you currently have here you can develop your own custom jobs and workflows within the sfc project directory. This is entirely up to you but designing a release job is a good idea. Ideally you should give access to the pypi project to LF so that we can place the credentials in Jenkins too. That way you won't have to hardcode your credentials in the job for Jenkins to push to PyPi and it will be safely stored in Jenkins config management.

Does this make sense?


 > Another question I would have is, if using the global-jjb job, how
 > do we make it run only on changes within the sfc-py dir (the
 > tox-dir) only?

Job doesn't currently support a "file-paths" option but we could add it to global-jjb [0]. Global-JJB is an open source project so we're happy to accept contributions here [1]. Or you can open a Jira [2] against global-jjb to submit the request.


[0] https://github.com/lfit/releng-global-jjb/blob/master/jjb/lf-python-jobs.yaml#L92-L133
[1] https://gerrit.linuxfoundation.org/infra/#/admin/projects/releng/global-jjb
[2] https://jira.linuxfoundation.org

Patch-set: 3
Label: Code-Review=0