Update patch set 3
Patch Set 3: Code-Review+1

Ready to merge. This patch does nothing until global-jjb is updated but it preps our repo for it and does not harm living in the repo.

We can tweak as necessary when global-jjb is released if API changes but I don't expect it will.

Patch-set: 3
Label: Code-Review=+1