Update patch set 14
Patch Set 14:

> Okay, looks good, thanks for putting up with my pedantic test
 > results expectations :)
 > 
 > Just send that last patchset that checks out master and I think
 > it's ready to merge.

hahaha your pedantic results expectations are nothing compared to my paranoia about things breaking because of bugs inserted by me.

Testing with master will not work because we will be missing the changes done in the scripts. For example, it will not find the URL variable since this patch is missing https://git.opendaylight.org/gerrit/#/c/68861/14/jjb/packaging/install-rpm.sh. So, all test will fail. I'll run them and check that they fail because of this

Patch-set: 14