Update patch set 12
Patch Set 12:

Hi Venkat,

I updated the patch to add the jjb configuration to add the openstack-pike packer job. This will allow you to upload the job yourself on the sandbox system in the future and generate your own image.

The important sections of the configuration is the "template" parameter which is the name of the json file for the project you want to build and the "platforms" parameter which is a "list" of supported platforms centos-7 or ubuntu-16.04. In this case I believe you're only using centos-7 so I only listed 1 platform.

When the job run completes look for a line like this:

==> openstack: Creating the image: ZZCI - CentOS 7 - devstack - 20180301-1004

Which is the name of the image that you can plug into your job for testing.

Hope this makes things easier for you.

[0] https://jenkins.opendaylight.org/sandbox/job/builder-packer-merge-centos-7-openstack-pike/

Patch-set: 12