Update patch set 12
Patch Set 12:

Forgot to mention in my last message a crucial step. When you click build make sure you set GERRIT_REFSPEC parameter to "refs/changes/19/66319/12"

Update the parameter as necessary to get the right patch. This is because since the patch is not yet merged we have to tell it to fetch the review so that it builds the right code.

Patch-set: 12