Update patch set 2
Patch Set 2:

Second thought, it is probably good to have this update here to ensure we get the latest code for the job, unless these branches don't change much and we should just update the vm images. If the branch is changing a lot then having it here is good, but if not the vm images is better.

Patch-set: 2