Update patch set 5
Patch Set 5:

(1 comment)

> Patch Set 5:
> 
> Not sure why this is failing in CI. Can anyone give me some hints?

this is a JJB failure. those are never easy to understand, but you can
see that it's the jjb-verify script that's failing if you dig through
the console logs.

to debug, I did this.

1) double check that I can "test" a packaging job (since that's the jobs
   this patch affects), but locally (on master) running this command:
   jenkins-jobs --conf ../jenkins.ini test ./ packaging-build-deb-nitrogen

   that passes and I get some XML output which is the actual job config

2) run the same command with this patch and you wont get the job config
   XML and if you scroll up through the output, you'll see a traceback.

3) started commenting out lines from this change. I noticed the line for
   test-ports-nofeature.sh was the problem. Then I noticed it was used
   in one other place of that .yaml, but used include-raw-escape.

4) change to use include-raw-escape and try 2) again. it passes.

I forget the gory details about include-raw vs include-raw-escape, but it
has something to do with variable expansion I think.

hope it helps.

Patch-set: 5
Label: Code-Review=0