Update patch set 1
Patch Set 1:

If we exit 0 everywhere then when a build failure does happen the script doesn't stop and keeps running and then the job might say it passed.


Maybe a better way to handle this is have the log pulling scripts run as a build publisher instead. In that case the builders will still fail the build and then when it moves on to the publisher section it will be grabbing relevant logs.

Patch-set: 1
Reviewer: Gerrit User 2759 <2759@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0