Update patch set 4
Patch Set 4:

> I would like feedback on this patch, but we will ned to merge this
 > other patches before hand, with manula verification:
 > https://git.opendaylight.org/gerrit/#/c/68874/
 > https://git.opendaylight.org/gerrit/#/c/69490/
 > https://git.opendaylight.org/gerrit/#/c/69492/
 > https://git.opendaylight.org/gerrit/#/c/69493/

You can push the job to the Jenkins sandbox. When running the job manually edit the "GERRIT_REFSPEC" parameter to point it to the patch you would like to test against. This should allow you to run the verify job against the patch in the Jenkins Sandbox.

Docs on how to push jobs to the sandbox is available here:

http://docs.opendaylight.org/projects/releng-builder/en/latest/jenkins.html#jenkins-sandbox

Hope this helps,
Thanh

Patch-set: 4