Update patch set 1
Patch Set 1:

as far as frequency goes, I had the same thought, we want to cat stuff earlier, but then I thought if the project really cared they would be more active, so just running weekly or so is likely more than enough. But maybe there are unmanaged projects which are a little more proactive. Again,though, these jobs should be there for the project to consume and not for releng to be off doing the work and keeping them happy.

Having the csit run on merge may not be so bad as long as it isn't frequent. I wasn't sure if we did a bump or something if all these merge jobs would bog things down

The other comment about opting in - I think the project should push the patches to do what they want as opposed to opting in via an email. This might really shake out projects who actually have a csit they care about.

Patch-set: 1
Reviewer: Gerrit User 773 <773@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=0