Check for 'recheck$' in comment triggers
It is best practice to ensure that we are checking for exact keywords
rather then if they appear anywhere in the comment message. This
prevents folks from accidently triggering builds as part of a
sentence.

global-jjb already standardizes on this format but releng/builder did
not pull it in yet and has caused confusion to folks as sometimes they
leave a comment after the trigger expecting all jobs to trigger. This
patch brings us inline with the best practice from global-jjb.

Change-Id: Iafc87e2b9c05727a2a309b9d9f78fa849702d262
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>