Convert os_plugins from csv to ssv
seems our deploy scripts are looping over the list
of openstack plugins that are enabled in each job.
in one place if one of the plugins was networking-sfc,
it would install a special CLI package so those
commands would be available and work. It looks like
the looping logic was expecting space separated,
but we are passing in comma separated lists. this
should fix that.

Change-Id: I0957b7f01a54948865aa942b4ecc6c40f69a5a07
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>