Create a queue for distro-test jobs
Considering all the distro-test job does is launches additional jobs
and having too many of them run in parallel makes it take longer for
any single distro-test job to return with results. Let's create a
new instance that we can use as a dummy queue for this job type
limiting only 1 of this job to ever run at a time.

Change-Id: Id3e94c1289f7ce89821ef9d53ddcb570f91f00b3
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>