Dont save non-existent tempest_output* files
the file(s) we are attempting to save for branches
other than queens don't exist anyway. At least
nothing is showing up in the current jobs.

This is a problem once we merge the change [0]
to manually install os-testr to get around our
problem of tempest.api.network tests not running
in branches before queens. The os-testr method
will create the ${DEVSTACK_TEMPEST_DIR}/${TESTREPO}/0
folder which makes this block of code execute
and then fail. In the end, our job runs find but
shows up as an abort.

Change-Id: Ie065944129a0cfa9ef06cd5b37d7274a691e33a1
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>