Separate test-plan for genius carbon-nitro streams
We are in the process of increasing test coverage for genius
functionalities. Most of the new features in genius have come
from Oxygen onwards. Hence, we want to have a default test plan
that covers the super-set of test suites, and a separate test plan
with basic suites that apply only to carbon and nitrogen.

Change-Id: If86b33b8bd72877ed46b179e8f32a2878cec22fb
Signed-off-by: Faseela K <faseela.k@ericsson.com>