Split multipatch into two phases, clone and build
This loops through all the patches and sets them up before starting a
build. This ensures that we fail fast if the patch specification is
incorrect, or if a patch can’t be checked out. It also paves the way
for better organisation of the build (automatic calculation of the
order...).

To make testing easier, this also adds a shebang and makes the script
executable.

Change-Id: I062741204c66df6f3b5805b24c63059c1df4177d
Signed-off-by: Stephen Kitt <skitt@redhat.com>