Determine build order in the job
This enhances the multipatch script so that it determines the build
order itself. It calculates effective POMs where necessary (so that we
only pull in POMs which are really used), pulls out the groups built
and used by each project, and feeds that to make to run the builds.

JIRA: RELENG-101
Change-Id: Ife878fb09338552466fbe17b6e616a7fe14a3996
Signed-off-by: Stephen Kitt <skitt@redhat.com>