Update patch set 1
Patch Set 1: Code-Review-2

can you please run our other non-hwvtep jobs in the sandbox with
this patch? just want to be careful as this will affect all of
our openstack jobs. probably, it's fine, but don't want to accidentally
break something.

btw, for every job you want to share, do a copy-logs: so it will
be around after the weekend purges that happen.

Patch-set: 1
Label: Code-Review=-2