Update patch set 2
Patch Set 2: Code-Review+1

> I thought the idea of this patch was to prevent duplicate image
 > names, which is why I suggested UUID in light of Thanh's doubt
 > that adding seconds would always be good enough.

Jamo, thanks for merging this change. 

if we are allowing images to be built on sandbox and prod, there is a possibility that jobs are run on the same time on both envs leading to images names clash when timestamped in mins, reduces the probability of a clash when timestamped in sec.millisec. the script look for image names and not uuid's for purging them.

uuid's can be used in lieu of a timestamp in the image name, but this requires us to update the scripts to search the JJB code with uuid and also use uuid's in the JJB code. uuid's are hard to know when it was built with doing a `openstack image show uuid`

Patch-set: 2
Reviewer: Gerrit User 6258 <6258@7fc14799-209e-464c-9743-7a06c2c21a81>