Update patch set 2
Patch Set 2:

If we only use seconds as original proposed there is a 1/60 chance of collision if 2 jobs start at the same time. Using ms we now have a 1/60000 chance of collision which is a little better but I think we're still betting on 2 simultaneous jobs having a slight delay between them.

If I understand correctly this issue came about because we pushed some packer jobs to the Jenkins Sandbox and it triggered via the monthly rebuild timer. If that's the case we're still betting on 2 jobs being queued at the exact same moment having a slight delay in processing.

Patch-set: 2