Update patch set 2
Patch Set 2: Code-Review+1

I'm fine with this if this is the direction the community wants to go. I do have some concern that disabling this will mask test issues that autorelease would normally catch. Since they are disabled we'll never know there is an issue but maybe sometimes ignorance is bliss? :)

Patch-set: 2