Update patch set 2
Patch Set 2:

> (1 comment)

That's a good idea but I wouldn't implement it as proposed. The proposal is asking us to double the amount of autorelease jobs that run and our infra just cannot handle that many autorelease jobs running simultaniously.

I think it would be better to investigate a way of running the SFT tests only, against the autorelease staged repo produced by the release job. This would be faster and doesn't require rebuilding autorelease twice for every release branch.

Patch-set: 2