Update patch set 10
Patch Set 10:

> > why not add this to tdd-wip instead of a separate job altogether?
 > > If separate, at least retain the tdd-wip suffix.
 > 
 > Do you mean add it to tdd-wip till failures are resolved, and then
 > move to nightly job?

There are no changes needed on test side for this, and it is just enabling the flag. So do we really need to make it part of tdd-wip. I need a job to be running nightly for this config, as it is our downstream default, and we have to make sure nothing is broken when we sync the code downstream one fine day and start using this mode

Patch-set: 10
Label: Verified=0