Avoid releasing projects whose name ends with 'r'
The release handling only looks for 'r' at the end of the patch name,
which also matches projects whose names end with 'r' if they are
included with no patch. This changes the release handling so that it
only activates if the patch name ends with 'r' and doesn't match the
project name.

Change-Id: I1f3045825e319b0a9f215913e5f0339a1f2c5b63
JIRA: RELENG-128
Signed-off-by: Stephen Kitt <skitt@redhat.com>