Update Gerrit trigger to use releng macro
There seems to be a new way to add triggers for Gerrit patch
submissions. We were defining a trigger manually ourselves in the
verify-rpm job, but now there's a shared trigger in releng-macros we can
use.

I noticed this because the packaging-verify-rpm-master job hasn't run in
quite a while, although it should have run on some incoming patches that
touched the packages/ dir. Hoping using this macro fixes the job.

Change-Id: Ib0cd9853ac621a7fe474cfaa058d0ec1bed117f8
Jira: INTPAK-179
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>