Update patch set 1
Patch Set 1:

I pushed this to the sandbox here:

https://jenkins.opendaylight.org/sandbox/view/packaging/job/packaging-verify-rpm-master/

And sent a new patchset to a test patch that makes a change to the packages/ dir:

https://git.opendaylight.org/gerrit/#/c/73490/2

But it doesn't look like the job in the sandbox picked up the Gerrit patch submitted event and triggered the verify-rpm job. I guess this could be because the fix doesn't work, or it could be because sandbox jobs prevented from triggering somehow. I'll send the tox verify job to the sandbox as well to test which it is.

Patch-set: 1