Update patch set 1
Patch Set 1: Code-Review+1

I pushed the packaging-tox-verify job to the sandbox and sent another patchset to my test change that updates the packages/ dir.

https://jenkins.opendaylight.org/sandbox/view/packaging/job/packaging-tox-verify-master/

https://jenkins.opendaylight.org/sandbox/view/packaging/job/packaging-verify-rpm-master/

https://git.opendaylight.org/gerrit/#/c/73490/3

Neither should-trigger-because-of-file-paths job triggered in the sandbox. So it seems I can't test this using the sandbox. Not sure how else to check, so +1.

Patch-set: 1
Label: Code-Review=+1