Split cluster csit and verify jobs
we want to run the normal controller cluster csit job
with -all- so we can catch any issues that might come
from conflicts in other projects, but we cannot install
restconf-noauth there because the -all- bucket of
features will also bring in non-noauth restconf and we
might hit conflicts.

but, our cluster verify job runs as -only- so we still
need to make sure to install restconf there.

Change-Id: I6c0e93989d821e6f9417e2f232fb8ce2f8637ff1
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>