Add odl-restconf to install-features
The -all- job that runs the normal csit will bring in the
compatible-with-all set which already includes odl-restconf.
However, the verify csit job is the -only- version which will
only load the features in install-features. Since the
odl-clustering-test-app feature does not include restconf
we need to have that in this list.

Change-Id: I6c0e93989d821e6f9417e2f232fb8ce2f8637ff1
Signed-off-by: Jamo Luhrsen <jluhrsen@redhat.com>