Add queens default
This patch adds the queens default in defaults.yaml
and changes the project.yamls to use that default in
their queens jobs. The default, though, is still pointing
to a pike image until the queens image is fixed as it
currently has issues with default ovs 2.9.0.

This patch makes it easy in the future to simply change
the value in the defaults.yaml for the project yamls
to pick it up when ready.

Change-Id: I641e74aafbd6c98e649da7966c3e4a222bf68b6c
Signed-off-by: Sam Hague <shague@redhat.com>