Update recheck trigger to match global-jjb
Upstream global-jjb is updating the 'recheck' trigger to be more
strict and support the 'reverify' keyword as well. This patch updates
the triggers to match in preparation for incoming upstream changes.

https://gerrit.linuxfoundation.org/infra/10705

In the past we removed support for "reverify" to make things more
simple but considering we want to keep consistency across LFN projects
we need to add it back as other LFN projects use reverify.

Change-Id: If941f9faf40dd9a21d03c9869a41c55724125a22
Signed-off-by: Thanh Ha <thanh.ha@linuxfoundation.org>